Moved FormSectionId type to shared types
