+++ frontend/src/features/data_entry/components/candidates_votes/CandidatesVotesForm.test.tsx
@@ -11,7 +11,6 @@ import {
 } from "@/testing/api-mocks";
 
 import { DataEntryProvider } from "../../stores/DataEntryProvider";
-import { DataEntryState } from "../../stores/types";
 import { errorWarningMocks, getDefaultFormSection, getEmptyDataEntryRequest } from "../../testing/mock-data";
 import {
   expectFieldsToBeInvalidAndToHaveAccessibleErrorMessage,
@@ -21,6 +20,7 @@ import {
   getCandidateFullNamesFromMockData,
   overrideServerClaimDataEntryResponse,
 } from "../../testing/test.utils";
+import { DataEntryState } from "../../types/types";
 import { CandidatesVotesForm } from "./CandidatesVotesForm";
 
 const defaultDataEntryState: DataEntryState = {
