+++ frontend/src/features/data_entry/components/candidates_votes/useCandidateVotes.ts
@@ -1,7 +1,7 @@
 import * as React from "react";
 
-import { SubmitCurrentFormOptions } from "../../stores/types";
 import { useDataEntryFormSection } from "../../stores/useDataEntryFormSection";
+import { SubmitCurrentFormOptions } from "../../types/types";
 import {
   CandidateVotesFormValues,
   CandidateVotesValues,
