+++ frontend/src/features/data_entry/stores/DataEntryContext.ts
@@ -1,5 +1,5 @@
 import { createContext } from "react";
 
-import { DataEntryStateAndActionsLoaded } from "./types";
+import { DataEntryStateAndActionsLoaded } from "../types/types";
 
 export const DataEntryContext = createContext<DataEntryStateAndActionsLoaded | null>(null);
