+++ frontend/src/types/types.ts
@@ -1,0 +1,6 @@
+export type FormSectionId =
+  | "recounted"
+  | "voters_votes_counts"
+  | "differences_counts"
+  | `political_group_votes_${number}`
+  | "save";
